Organize loaders module
- Group into submodules:
    - `loaders/schema`: for the ArtifactsLoader and GraphQL hookup
    - `loaders/resources`: for GraphQL logic (incl. queries/mutations)
    - `loaders/commands`: for dedicated compile/migrate/etc. workflows

- Organize individual GraphQL query/mutation definitions into individual
  resource submodules and use *.graphql.ts suffix

- Rename AddCompilation -> AddCompilations
