--- packages/db/src/loaders/queries/index.ts
@@ -1,12 +1,0 @@
-export { AddBytecodes } from "./bytecodes";
-export { AddSources } from "./sources";
-export { AddCompilation } from "./compilation";
-export { AddContracts } from "./contracts";
-export { AddContractInstances } from "./contractInstances";
-export { AddNameRecords } from "./nameRecords";
-export { AddNetworks } from "./networks";
-export {
-  AddProjects,
-  ResolveProjectName,
-  AssignProjectNames
-} from "./projects";
