+++ packages/db/src/loaders/resources/bytecodes/index.ts
@@ -1,0 +1,2 @@
+import { AddBytecodes } from "./add.graphql";
+export { AddBytecodes };
