+++ packages/db/src/loaders/resources/compilations/index.ts
@@ -1,0 +1,2 @@
+import { AddCompilations } from "./add.graphql";
+export { AddCompilations };
