+++ packages/db/src/loaders/resources/contractInstances/index.ts
@@ -1,0 +1,1 @@
+export { AddContractInstances } from "./add.graphql";
