+++ packages/db/src/loaders/resources/nameRecords/index.ts
@@ -1,0 +1,2 @@
+import { AddNameRecords } from "./add.graphql";
+export { AddNameRecords };
