+++ packages/db/src/loaders/resources/projects/index.ts
@@ -1,0 +1,4 @@
+import { AddProjects } from "./add.graphql";
+import { AssignProjectNames } from "./assign.graphql";
+import { ResolveProjectName } from "./resolve.graphql";
+export { AddProjects, AssignProjectNames, ResolveProjectName };
