+++ packages/db/src/loaders/resources/sources/index.ts
@@ -1,0 +1,2 @@
+import { AddSources } from "./add.graphql";
+export { AddSources };
