+++ packages/db/src/loaders/schema/test/compilationSources/MagicSquare.sol
