+++ packages/db/src/loaders/schema/test/compilationSources/VyperStorage.vy
